update: migration

+++ database/migrations/2025_02_11_225907_table_planteles.php
@@ -19,7 +19,7 @@ return new class extends Migration
                 $table->string('direccion')->nullable();
                 $table->string('telefono')->nullable();
                 $table->string('correo')->unique()->nullable();
-                $table->integer('estado_id'); // Ensure this matches the data type of idestado in estados table
+                $table->unsignedBigInteger('estado_id'); // Ensure this matches the data type of idestado in estados table
                 $table->foreign('estado_id')->references('idestado')->on('estados')->onDelete('cascade');
                 $table->timestamps();
             });
